moved InferSchema to read package
